cache, settings, ymal, ...

+++ src/marknotes/plugins/task/optimize/clear.php
@@ -12,25 +12,38 @@ class Clear
 {
 	public static function clearSession()
 	{
-		$aeCache = \MarkNotes\Cache::getInstance();
 		$aeSession = \MarkNotes\Session::getInstance();
-
-		/*<!-- build:debug -->*/
 		$aeSettings = \MarkNotes\Settings::getInstance();
-		if ($aeSettings->getDebugMode()) {
-			$aeDebug = \MarkNotes\Debug::getInstance();
-			$aeDebug->log("Clear","debug");
+
+		if (boolval($aeSession->get('authenticated', 0))) {
+
+			/*<!-- build:debug -->*/
+			if ($aeSettings->getDebugMode()) {
+				$aeDebug = \MarkNotes\Debug::getInstance();
+				$aeDebug->log("Clear","debug");
+			}
+			/*<!-- endbuild -->*/
+
+			// Clear the cache
+			$aeCache = \MarkNotes\Cache::getInstance();
+			$aeCache->clear();
+
+			// When the task is 'clear', just clear the session
+			$aeSession->destroy();
+
+			header('Content-Type: application/json');
+			echo json_encode(array('status' => '1'));
+		} else {
+			// The user isn't logged in, he can't modify settings
+
+			header('Content-Type: application/json');
+			echo json_encode(
+				array(
+					'status'=>0,
+					'message'=>$aeSettings->getText('not_authenticated')
+					)
+				);
 		}
-		/*<!-- endbuild -->*/
-
-		// Clear the cache
-		$aeCache->clear();
-
-		// When the task is 'clear', just clear the session
-		$aeSession->destroy();
-
-		header('Content-Type: application/json');
-		echo json_encode(array('status' => '1'));
 
 		return true;
 	}
